server/requirements.txt
crewai==0.28.8
Flask==2.3.2
Flask_Cors==4.0.0
gradio_client==0.8.0
langchain==0.1.17
langchain_core==0.1.50
langchain_openai==0.0.5
Pillow==9.4.0
pydub==0.25.1
python-dotenv==1.0.0
redis==4.4.2
Requests==2.31.0
SQLAlchemy==2.0.29
transformers==4.40.2
psycopg2==2.9.2
replicate==0.15.4